# api/openapi.yaml
openapi: 3.0.3
info:
  title: Go Base Microservice
  description: A minimal Go microservice template with PostgreSQL, Kafka, and Kubernetes health endpoints
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  schemas:
...
    LogLevel:
      type: object
      properties:
        level:
          type: string
          enum: [debug, info, warn, error]
          example: info
    LogLevelRequest:
      type: object
      required: [level]
      properties:
        level:
          type: string
          enum: [debug, info, warn, error]
          example: debug
    LogLevelResponse:
      type: object
      properties:
        level:
          type: string
          example: debug
        message:
          type: string
          example: Log level updated successfully
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
...
paths:
...
  /api/v1/admin/log-level:
    get:
      summary: Get current log level
      description: Returns the current log level of the application
      tags: [Admin]
      operationId: getLogLevel
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogLevel'
    put:
      summary: Change log level
      description: Dynamically changes the log level of the application
      tags: [Admin]
      operationId: updateLogLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogLevelRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogLevelResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "invalid log level: trace"
...
  /api/v1/echo:
    post:
      summary: Echo endpoint
      description: Echoes back the request body
      tags: [API]
      operationId: postEcho
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              message: "Hello World"
              data:
                "key": "value"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
              example:
                message: "Hello World"
                data:
                  "key": "value"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid JSON body"